Extracted errorsPlugin test
